requirements.txt
alembic==1.8.1
anyio==3.6.1
astroid==2.12.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.1
dill==0.3.5.1
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
freezegun==1.2.2
greenlet==1.1.3
h11==0.13.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.2.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
pytest-freezegun==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
websockets==10.3
wrapt==1.14.1
zipp==3.8.1
